LayoutTests/fast/css/parsing-text-rendering-expected.txt

This tests checks that all of the input values for background-repeat parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("text-rendering: auto;") is "auto"
PASS test("text-rendering: optimizeSpeed;") is "optimizeSpeed"
PASS test("text-rendering: optimizeLegibility;") is "optimizeLegibility"
PASS test("text-rendering: geometricPrecision;") is "geometricPrecision"
PASS test("text-rendering: auto auto;") is ""
PASS test("text-rendering: optimizeCoconuts;") is ""
PASS test("text-rendering: 15;") is ""
PASS successfullyParsed is true

TEST COMPLETE
